fix(robustness): NaN panic + overflow + unwrap + number parsing
- batch_scheduler: partial_cmp().unwrap() → unwrap_or(Equal) 避免 NaN panic
- batch_scheduler: current_batch_size Relaxed → Acquire/Release 保证跨线程可见性
- candle_engine: check_memory_pressure u64 乘法溢出 → u128 中间值
- candle_engine: estimate_parameter_count/get_hidden_size 添加 TODO 标注局限性
- memory_paging: unwrap() → ok_or_else(LayerNotFound) 一致错误处理
- amd: extract_number_from_line 提取首个连续数字序列而非拼接全部数字

+++ src/engine/candle_engine.rs
@@ -524,7 +524,8 @@ impl CandleEngine {
         if let Some(ref monitor) = self.memory_monitor {
             let stats = monitor.get_memory_stats().await;
             let usage_percent = if stats.total_bytes > 0 {
-                (stats.current_bytes * 100) / stats.total_bytes
+                // 使用 u128 中间值避免 current_bytes * 100 溢出 u64
+                ((stats.current_bytes as u128 * 100) / stats.total_bytes as u128) as u64
             } else {
                 0
             };
@@ -943,18 +944,19 @@ impl CandleEngine {
 
     /// 估算模型参数数量
     fn estimate_parameter_count(&self) -> u64 {
-        // 基于 BERT-Base 的估算（约 110M 参数）
+        // TODO: 从模型权重元数据或 config 读取实际参数数量，当前仅适用于 base 变体
         match &self.model_architecture {
             ModelArchitecture::Bert => 110_000_000,
-            ModelArchitecture::XlmRoberta => 270_000_000, // XLM-RoBERTa base 约为 270M
+            ModelArchitecture::XlmRoberta => 270_000_000,
         }
     }
 
     /// 获取隐藏层大小
     fn get_hidden_size(&self) -> usize {
+        // TODO: 从 config.hidden_size 读取实际值，当前仅适用于 base 变体（BERT-Base/XLM-RoBERTa-Base 均为 768）
         match &self.model_architecture {
-            ModelArchitecture::Bert => 768,       // BERT-Base
-            ModelArchitecture::XlmRoberta => 768, // XLM-RoBERTa-Base
+            ModelArchitecture::Bert => 768,
+            ModelArchitecture::XlmRoberta => 768,
         }
     }
